sample step should be i32 not i64

+++ gremlin-client/src/process/traversal/builder.rs
@@ -485,7 +485,7 @@ impl TraversalBuilder {
         self
     }
 
-    pub fn sample(mut self, step: i64) -> Self {
+    pub fn sample(mut self, step: i32) -> Self {
         self.bytecode
             .add_step(String::from("sample"), vec![step.into()]);
         self
